refactor: extract model crate, implements fetch-by-alias
