+++ Cargo.toml
@@ -35,6 +35,7 @@ members = [
     "collector/nvd",
     "v11y/api",
     "v11y/client",
+    "v11y/model",
     "v11y/v11y",
     "integration-tests",
     "trust",
@@ -81,6 +82,7 @@ default-members = [
     "collector/nvd",
     "v11y/api",
     "v11y/client",
+    "v11y/model",
     "v11y/v11y",
     "trust",
     "exporter",
