+++ spog/model/Cargo.toml
@@ -15,6 +15,7 @@ time = { version = "0.3" }
 url = { version = "2", features = ["serde"] }
 utoipa = { version = "3" }
 vexination-model = { path = "../../vexination/model" }
+v11y-model = { path = "../../v11y/model" }
 
 # required by ToSchema utopia
 serde_json = "1"
