+++ v11y/api/Cargo.toml
@@ -7,7 +7,7 @@ edition = "2021"
 
 [dependencies]
 actix-web = "4"
-v11y-client = { path = "../client" }
+v11y-model = { path = "../model" }
 trustification-auth = { path = "../../auth", features = ["actix", "swagger"] }
 trustification-infrastructure = { path = "../../infrastructure" }
 clap = { version = "4", features = ["derive"] }
