+++ v11y/api/src/server/vulnerability.rs
@@ -3,7 +3,7 @@ use trustification_auth::authenticator::user::UserInformation;
 use trustification_auth::authorizer::Authorizer;
 use trustification_auth::Permission;
 
-use v11y_client::Vulnerability;
+use v11y_model::Vulnerability;
 
 use crate::db::{DbError, GetBy};
 use crate::server::Error;
@@ -68,7 +68,7 @@ mod test {
     use crate::AppState;
     use actix_web::{test, test::TestRequest, web, App};
     use std::sync::Arc;
-    use v11y_client::Vulnerability;
+    use v11y_model::Vulnerability;
 
     #[actix_web::test]
     async fn get_none() -> Result<(), anyhow::Error> {
