+++ v11y/client/Cargo.toml
@@ -4,18 +4,13 @@ version = "0.1.0"
 edition = "2021"
 
 [dependencies]
-#actix-web = "4"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0.68"
 tokio = { version = "1.0", features = ["full"] }
 log = "0.4"
-humantime = "2.1.0"
-humantime-serde = "1.1.1"
-#clap = { version = "4", features = ["derive"] }
 reqwest = "0.11.18"
 anyhow = "1"
-chrono = { version = "0.4.26", features = ["serde"] }
-utoipa = { version = "3", features = ["actix_extras"] }
+v11y-model = { path = "../model" }
 
 trustification-auth = { path = "../../auth" }
